Detect non-build uses of ocamlbuild, and add an unconstrained dependency
The previous bulk-update work only tried to detect packages using
ocamlbuild as a build system; not packages distributing ocamlbuild
plugins, or any other library or executable linked with OCamlbuild. In
this latter (minority) use-case, adding a build-time dependency only
is insufficient.

Gabriel Radanne suggests detecting ocamlbuild plugins by looking at
the string "Ocamlbuild_plugin" appearing in source code. This base
idea needed to be refined quite a bit to remove a fair number of false
positives (see the details in the commit), but we are now able to
detect 65 packages that have a non-build-time dependency over ocamlbuild.

I manually checked the list of updated packages, so I know there are
no false positives (of course re-running the script later may
introduce some); on the other hand, it is possible that some package
that *fail* this test actually depend on ocamlbuild, so it's still
a good idea for plugin distributors to check their packages.