ocamlfind does not depend on ncurses
There have been several installation issues with ocamlfind recently due
to the fact that conf-ncurses uses pkg-config to check installability,
which is only valid under recent ncurses versions. But in fact ocamlfind
does not depend on ncurses at all (I just checked with Gerd Stolpmann),
so there is no reason for the dependency in the first place.

fixes #5880