fix build of coq+coqide 8.5.0 under 4.03.0
For coq 8.5.1, a single patch is necessary to fix the build under 4.03.0.

For coq 8.5.0, an additional patch is necessary (it is a patch from
Coq developers upstream,
  https://github.com/coq/coq/commit/b98e4857a13a4014c65882af5321ebdb09f41890
that happened after the 8.5.0 release but compiles as-is against 8.5)

For typical software I would just add a version bound (< 4.03.0) on
8.5.0 and encourage users to use 8.5.1 instead, but Coq is peculiar is
that its compatibility breakage are frequent enough that users tend to
need very specific versions. The more old-coq-versions we can get to
build under 4.03.0, the better.

Note: the commit sha1 that occurs in files/ephemeron-rename.patch
( 2bd33495267e1d5c7eae507c25d4d7a49b3b61f4 ) is not the same as the
upstream one ( b98e4857a13a4014c65882af5321ebdb09f41890 ) as it
corresponds to the versions that I re-applied on top of the pristine
8.5.0 release -- there are minor changes that were automatically
adjusted by `git cherry-pick`.