conf-ncurses: do not enforce the use of pkg-config
Some systems, such as OpenSUSE, do not install pkg-config descriptions
reliably, and so this package unnecessarily fails despite ncurses
being installed fine.

This PR disables the enforced `pkg-config ncurses` invocation but
retains the depext hints, and adds an OpenSUSE depext.  With this,
installation works fine on OpenSUSE 42.1 and this user should be
happier: http://lambda.jstolarek.com/2016/05/installing-ocaml-under-opensuse-11-4-or-the-compilation-of-conf-ncurses-failed/