extlib, extlib-compat: work on 4.04+beta2, set weaker bound < 4.05
The dependency "< 4.04" was set in #7164 and #7173 because extlib was
broken by the string/bytes-primitives name change. This change has
been reverted from 4.04, precisely to improve compatibility, so these
extlib versions are compatible with 4.04 again.

(As far as I can tell the discussion in ygrek/ocaml-extlib#37 is now
irrelevant due to the revert of the backward-incompatible change:
there is no extlib change required for 4.04 compatibility, so there is
no specific need for a new extlib release.)

Having extlib installable is important as it appears in dependency
chains. For example, opkg <- opam-lib <- cudf <- extlib.