conf-binutils: defer binutils installation
During the binutils installation there is a warning, that
it may broke installation of other packages. As practice
shows, this is indeed true. See #563.

Indeed, the binutils package installs the offending ansidecl.h file,
that later breaks at least mingw32-binutils. It might be better to drop
this package at all, but in this case we will loose the support for
darwin files (for the objdump package).

If later we will estimate that the amount of cons outweights the amount
of pros, provided by the bap-objdump package, then we can just drop this
package from the core bap installation and make it optional.