[BAP] fix bap-llvm build on Travis CI
Ubuntu image on Travis is specially alterated, so that the `llvm-config`
is relinked from the default (and installed 3.4) into
`llvm-config-3.5`, that is not supported by BAP.

All dependencies are installed correctly by the depext (including
llvm-3.4), but due to some hardcoded specifics of travis, the
`llvm-config` file still points to 3.5 that we even didn't install. It
looks like, that travis team didn't use the `update-alternative` mechanism
to choose a particular version of llvm, but did some hardcode linking.

To prevent the failures, the following is proposed:

1. new `conf-env-llvm` package is added that will make the Travis
environment available as opam variables

2. use `conf-env-llvm` package to specialize configuration on Travis.