Do not add -L/usr/lib to camlzip.cm{x}a on OSX
This breaks when trying to use libffi (and thus ctypes), because
OSX has a old version of libffi in /usr/lib and we rely to the
version installed by homebrew.

Due to https://caml.inria.fr/mantis/view.php?id=7150 we don't have
any control on the order of -L arguments passed to the linker, so
it's just better to not add one.